requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
cssselect2==0.8.0
diff-match-patch==20241021
Django==5.1.6
django-debug-toolbar==5.0.1
django-import-export==4.3.5
django-widget-tweaks==1.5.0
fonttools==4.60.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
persiantools==5.2.0
pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-decouple==3.8
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
urllib3==2.3.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
